requirements.txt
asgiref==3.8.1
click==7.1.2
click-didyoumean==0.3.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
